deprecate [advancefrom]
we changed also its implementation to use [modify_unit_type]
instead. This simplifies the implementation.

Now it also can no longer cause OOS if used recklessly.
But the new implementation also might not cover 100% of
its previous usecases, in particular this means that
[advancefrom] now only effects [campaign] within the same
addon. This might sound like it breaks compatability for
addons that have their units in a differnt addon and used
[advancefrom] but most of those would have casued OOS
before, so i think this actually an improvement.