gui2/addon_manager: Re-read add-on versions after each update batch
The Update All codepath is unique in that it doesn't trigger a re-read
of add-on versions until the very end, once the full run has completed.
The problem with this is that install_addon_with_checks() may update
dependencies of individual add-ons in the process, resulting in
redundant downloads of any affected add-ons if the player isn't paying
attention.

Address this by triggering a re-read of add-on versions after every
single update batch that results in a download.

Closes #3273.