add student variable to ability formulas
This can be different from 'other' in case of a
an ability that effects the enemy, where 'other'
should be the unit that receives the effect
(here: the enemy), studend is the unit that
matches the [affect_adjacent], and the unnamed
unit is the unit that has the ability.

I think it'd be nice to have even more variables
available so that for example the effect could
depend on the opponent even when it doesn't
effect the opponent.