add lua weapon:active_specials function
This function can calculate the active specials,
(including abilities) with a given tagname. It
can return the wml tables for each active
special, calculate the value for an ability
with a value or just be used to check whether a
special with the given tagbane is active at all.

It takes a table with optional parameters at single
argument and returns a table, with the calculated
results.