add attacker and defender variables in ability formulas
Now formulas like "add=(attacker.level - defender.level)"
are possible in weapon specials that allow the author to compute
the abilities value based on the opponent. The main reason why i
choose for attacker/defender instead of self/opponent was thats
is easier to implement. (but using the self/student variable is
also possible to get a 'opponent' variable shoudl that be needed)
Also it is less ambiguous than 'self/opponent' in the case of
weapon specials that come from leadership-like abilities.