Add comments about minimap.cpp's use of "team colors"
Maybe the "team_rgb_range" should be renamed as "named_color_range", because
it's used for everything defined by the WML `[color_range]` tag. In addition to
the ones normally used for teams (`red`, `lightblue`, etc) it has the minimap
colors for `reef`, `cave`, etc.

Separating the team colors from the terrain colors probably shouldn't be done.
Looking from the opposite perspective, using the terrain colors as team colors
has been discussed on the forums and found to be a useful feature:

* https://r.wesnoth.org/p627592
* https://r.wesnoth.org/p627588