add World Conquest II to mainline
For mainline we will just call it "World Conquest"

World Conquest II is a coop multiplayer campaign
of 5 scenarios for 1,2 or 3 players.
It features highly randomized maps for
replayability. Also custom item and training
systems. The majority of the campaign is
implemented in lua for better preformance.

It also contains a custom era to be used together
woth the campaign, since the default era is
quite unbalanced for this coop gamemode against
the ai. It for example makes sure that every
faction has a leader.

This is one of the most popular MP addons on
wesnoth 1.14, 1.12 and older.