bi: Full refactoring of the clipboard report formatting logic
This makes all sections use a formatter class to deal with the issue of
keeping the first column's width consistent within each section (but not
across all sections), so that we don't run into inconveniences later in
the future if any section's entry labels change. It also allows us to
change the list formatting entirely in a single place rather than
multiple places if we ever decide to do so.

This also replaces CVideo::video_settings_report() with a similar but
slightly more detailed report using the new bi list formatter class to
make it look consistent with the other sections, implemented in the bi
module instead of CVideo.