Make forced --wconsole logic easier to read and extend
Instead of a massive wall of chained ORs, have two lists with qualifying
switches, one for those that don't take arguments and another for those
that do so. The latter are expected to have the argument past the '='
character (substituting the unwieldy hardcoded compare() method calls
with substr() comparisons) or in a subsequent command line argument that
we can ignore at this point.

Also added --validate-addon, -h and -v to the list of applicable
switches.

There's a lambda expression that needs to be redone on every loop but
who cares, this is only the command line argument scan phase anyway.