game_config_manger refactoring
This commit prepares a patch to disable unused
addons during a game. We try to reduce codes
that change the game_config config object.

In particular this removes a code that changes
the order of terrain graphics tag. This could
in theory break code that relied on that order
but afaik the terrain_graphics tags offers a
`precedence` that allows users to specify the
order in which terrain graphics are applied.
so no additional compatabiltiy code is needed.