gui2/unit_create: Use variation_name before the variation's ut name
Forgot that variation_name has an actual meaning when paired with
variation_id as opposed to it just being a deprecated form of
variation_id. This makes it so it's favoured by the UI over a
potentially overridden type name when present.

This makes it so the mainline Walking Corpse and Soulless units have
their translatable variation names properly displayed.