UtBS: Make "quenoth" a separate race to "elf" (fixes #4876)
This is motivated by reorganising the help pages to split the Quenoth from the
"pointy ears, pale skin" help text in data/core/units.cfg. It also means that
there's no reason to [hide_help] for the default era elvish units. The new help
text is just a short placeholder for now.

The new race has [race]help_taxonomy=elf.

If the player loads a mid-campaign save then their recalls and heroes will have
race=elf while new recruits have race=quenoth.  For that reason, I've changed
the scenarios' event triggers to recognise both races.

When levelling up, units with race=elf may change to race=quenoth.

Opening the help for a unit will show the help as if the unit has race=quenoth,
even when right-clicking on a race=elf unit. This is good.

For a few of these, I think another race might be better - they're converted to
Quenoth here, as they were Elves before. The Corrupted Elf is should probably be
considered undead, and the Divine Avatar should possibly be something else, more
like an elemental than a living creature. That those two are currently included
in the elves does have the side-effect of hiding a plot spoiler about the Dark
Assassin who is also included in the elves (the spoiler being that this one is
correct, he really is an elf).