Avoid prematurely generating bad color range/palette definitions
Previously it was possible for the game to try to coerce any named color
range/palette that hasn't been defined by WML yet into a range/palette
with RGB #000000 as its first entry, since strtol() (used by the
color_t::from_hex_string() function) will return 0 if the input is not a
valid hexadecimal number.

Now, normally this condition was *not* reached because
color_t::from_hex_string() throws std::invalid_argument if the input
isn't exactly 6 chars (as in char, not Unicode characters) long. The
problem is, in English, "yellow" is exactly 6 chars long.

The existing code suggests that we should be throwing in this case, and
this is exactly what was already being done in the general case, so just
do that immediately when:

 1. A named color range doesn't exist and it doesn't have at least one
    comma.
 2. A named color palette doesn't exist and it has characters that
    aren't hexadecimal digits or commas.

Reasoning:

 1. While color_range's type definition in the engine suggests that only
    the first colour value is needed, this doesn't make much sense from
    the WML API standpoint so we might as well be stricter than the
    color_range constructor here.

 2. As for palettes, there's a possibility that someone may want to
    perform a palette substitution with exactly one colour (e.g.
    `~PAL(FF00FF>000000)`), so we just make sure that all characters in
    the "name" are hexadecimal digits. This means that using "DEADED" as
    a name might be a bad idea if the palette somehow finds itself being
    used before being defined (it will effectively create a new palette
    with #DEADED as its only entry) but it's either that or making the
    single colour case surprisingly clunky to write.

Ideally, we should be resetting colour info entirely (the
game_config::reset_color_info() function) every time we enter a game,
but for *some* reason that's not done at the moment. It would also make
more sense to allow color_range and color_palette definitions to be
overridden under certain circumstances, but currently that's not the
case (the first defined element with a given id remains a constant for
the entire game session) and it's also more work to change that.

(Wow, I'm doing a terrible job at selling this commit, aren't I?)