add unit.active_attack lua api
this function can be queried to get the active
attack of a unit during an attack.
this is is particular more reliable than using
the attack wml from the event context since that
one won't have the correct attack_context set
when creating an attack object from it.

See also related issue #3317