Help dealing with losses: remove todo for linking to the forums
This simply removes the todo. Conclusions from trying to add a link:

* Any link should be to something more specific than the "Mainline Campaigns
Feedback" forum, which is a set of sub-forums that may or may not have any
recent and-or useful information for the particular scenario that the user is
having trouble with. For replays in particular, some (most?) don't have replays
that have been uploaded to the feedback threads, and given Wesnoth's
incompatibility with previous stable branches the replays that are there might
not be usable.

*`The forums are English, so might want a note in all the non-English
translations that isn't in the English text.