Macros: fix leaked variable in {HEAL_UNIT}
[heal_unit] set's as side-effect an variable, which contains the heal amount
of one of the healed units.
It's more likely that this will just add a variable with authors not being aware
to clean it afterwards, than the case being that someone want's this behaviour.

I assume that this happens is an unintended side-effect from former code refactoring,
the original implementation of this macro was not doing this.

[ci skip]