Save-loading: remove "under discussion", add hint about accepting losses
This removes some subtle todo text that didn't have a explicit "TODO" in
the text, although it was tracked as part of #4280. The discussion ended
without a conclusion about whether those statements were correct for all
mainline campaigns.

The existing text in this help topic represented paragraphs by having
translatable strings that start with whitespace (two newlines), as described in
issue #4480. As this commit already needed to change all but two of the
paragraphs, I've switched them all to using non-translatable paragraph
separators.  Even though the whole save-loading topic is new in 1.15, it has
already been translated in some languages, and making this consistent means two
unnecessary fuzzy strings. Sorry to the translators, but I think it's
worthwhile.