WC: First pot update
Because the first pot update automatically generates translations for
the English locales from the source, en@shaw and en_GB get automatically
populated with msgstrs identical to their respective msgids. As far as I
understand from the discussions around the inclusion of WoV, this is
expected behaviour for en_GB, but undesirable for en@shaw (different
language script), so I've manually blanked out the latter in this
commit.

(Allowing CI builds for this one to ensure translations can be built
before merging the PR.)