TSG: move objectives to prestart
In some cases, players can see the obectives already earlier
by pressing CTRL + J. E.g. when the game scrolls to a unit
for a [message]. If the objectives aren't set at that time,
then the default objectives are displayed (kill all enemy leaders).

Because the code uses boolean_equals instead of equals it's not
anymore dependent of other variables being set beforehand.

[ci skip]