git-subtree like tool for exporting campaigns
As with git-subtree, it creates reproducible SHA1 ids for the new branch, the
history of An Orcish Incursion will appear as:

* master branch (pre-removal) 2637115bd92334444f1e03d9387266a92ae478da
* master branch (removal) 3d4a4b06 and 8deac9bb
* 1.14 branch 6b5ccb5fd12e7bb4795b25d1aa10d6f0c88a7490

Wesnoth's po/<textdomain> folder is copied into the subtree as "translations".
If there already was a "translations" in the subtree, it's renamed to
"umc_translations_dir".

The campaign will still need a few fixups for paths in _main.cfg, and
the scenarios' map_data or map_file attributes. The map_file paths
were made more portable in a5a1fec2, but this was after AOI's removal.

Many thanks to Lord Knightmare for porting AOI to 1.15 without this script,
which provided a reference to check the script against. LK's 1.0.0 matches
this script's output from the 1.14 branch, including the fix in b9f93cf that
was applied to 1.14 after the removal in master, and with the rank (but not the
"Rookie") from cc01ccae.