Add 'Delay Advancements [modification]
Depsite being a [modification] this is actuall implemented
in the C++ code the [modification] just sets a flag.

This is meant as an alterntive to the preset advancement
type of mod, i want to play around with it a bit in 1.15
maybe we will remove it later.

The problbme this is supposed to fix is that advancemnts
are done randomly during the enemeies turn. It has a few
advantages over the preset advancemnt approach:
1) It can easily handle amlas and normal advancement
2) It doesn't need special code to handle the case that
   A unit advances multiple times.
3) It doesn't break in case that other wml code changed
   The advancements  of a unit after the preselected advancement was chosen
4) The user never needs to think about an advancement of
   a unit that might not even advance

It also has a disadvantage: it changes the rules of the
game quite a bit, in partiucar if the units heals form
an adcancement (which us usally the case), since now
with this 'healing the unit on advancement by retaliation'
during the enemies turn can no longer happen. I still
think its wirth to think about this and test it though.