Update the docs for ":droid full" in the help browser
This command and the documentation provided by ":help droid" were updated
in efd808ea842ff88174cc2e953c8b871b8c41dae0, but we missed that this command
is also documented in the help browser's (Commands, General Commands) page.

The docs here are more verbose than in ":help", so the text is newly written
instead of being copied from the existing location.