Another go at refactoring distributor.cpp (builds and runs successfully)
This is part-way though a refactor, but I'm not committing myself to completing
it; publishing for the comments and discussion.

This solves PR #5198's compile errors about undefined references, but the
solution here is to expand the templates into the multiple hardcoded copies
block starting with `switch(events_.sdl_button_down_event)`.