Move argv repackaging logic out of wesnoth.cpp
This includes all the Win32-specific code dealing with the retrieval of
the UTF-16 version of the command line and subsequent parsing into an
array of UTF-8 strings.

The latter code is unchanged except for the function names and one local
variable originally named `is_excaped` instead of `is_escaped`.

(Note to self: ask gfgtdf if he was aware of the existence of
CommandLineToArgvW() back when he wrote the Win32 command line retrieval
logic.)