WIP: trackplacer support for IntroWML support for text labels on maps
Add support in tmx_trackplacer and trackviewer, for both .tmx and .cfg formats.
For translation purposes, the text gets "map_label^" prepended to it.

WIP: trackviewer doesn't have the correct y-coordinate for the label yet (but the
x is correct, it anchors to the center of the text).

WIP: check for todos in the code, there's one in tmxfileformat.py to hide an
error about the empty journey when reading .cfg files that the new code writes.

The C++ implementation would allow labels to be revealed in the middle of a
journey, however at the moment trackplacer only creates them as a separate
track, intended to be shown at the start at the same time as the background
map.