WIP: Add IntroWML support for text labels on maps
This adds a `[story][part][image]label=` attribute, allowing the `[image]` tag
to draw either images, text, or both. Putting both functionalities on the same
tag has tradeoffs - the tag name is now a bit inaccurate, however it allows
having landmarks on the map such as

    [image]
		x,y=...,...
		file="map_icons/dwarven_door.png"
		label="map_label^Kal Kartha"
	[/image]

For the label, the x and y coordinates aren't the exact placement of the text,
instead they're the location of the place that the text should label.

This only adds the engine support, it lacks the integration into trackplacer.