campaignd: Add server config, dir and port command line options
This changes startup behaviour slightly so that campaignd ensures that
the directory containing the config file exists before doing anything
else. Normally this wouldn't be noticeable since the default behaviour
continues to be to use ./server.cfg (pwd should exist, right?), but it's
important in the event that the user passes -c with a nonexistent path.

We do still create the server.cfg file later anyway, and it's still
possible that the directory containing it disappears in between only to
be forcefully created by ostream_file() -- but who cares, this is all
just meant to be an early reality check to avoid user mistakes.