addon/validation: Add add-on validation status messages
This implements a status message table for all currently possible add-on
validation statuses.

The idea is to aid making campaignd's side of the validation cleaner and
less data-laden, and to provide a path for implementing translation of
these messages on the client side without running into issues where
minor changes to strings on the server side forcefully fuzzy them until
the next release.

This is not used yet since the relevant code is part of a major
refactoring of the [upload] request handler.