campaignd: Whole-sale refactor of [upload] handler
This splits the entirety of the pre-upload validation checks at the
start of campaignd::server::handle_upload() into a new validate_addon()
method, which delivers information about the validation outcome through
its own interface, including the ADDON_CHECK_STATUS enum type.

It also largely changes code to avoid multiple invariant WML attribute
reads that, shifts further away from the 'campaign' term internally, and
more importantly, avoids making extra copies of the WML payload (which
can be as large as 100 MB in production) by modifying the request body
directly if necessary. This last part is very much not elegant, but it
will do for now.

Some code has also been reordered, reformatted, or replaced with
equivalent syntax for clarity.

There's still some work to be done here, especially with regards to
delta uploads. At least the handler itself should be more
straightforward and readable now. Additionally, this opens up the
possibility of implementing a mechanism to "pre-validate" an add-on
before uploading the full thing. The jury's still out on whether this
would be a good use of my time or not, though.

Also, as part of this commit, the send_error() and send_message()
methods (and all of their overload) escape their input for simple_wml
themselves instead of leaving it up to the caller to figure out and sort
out the most ridiculous API design decision in the entire history of
Wesnoth.