store addon versions in version_info
The recent campaignd rework allowed addons to provide multiple versions. The local version_info will now store all avaible version names in descending order and also keep the current version which will be able to be selected from a menu or may be used to store info for local purposes.

The versions set is guaranteed to have at least one entry, which is current_version.