campaignd: Avoid using config::remove_children() during loops on ranges
There are no obvious guarantees made with regards to the validity of
iterators returned by config::child_range() after a deletion. As far as
I can tell, it's very likely that holding an iterator to an element,
then deleting that element, then advancing that iterator could result in
UB. It could even result in items being skipped.